Fixed circular include problems for Acceleration calculator
