+++ Race/Simulation/AccelerationCalculator.h
@@ -1,7 +1,8 @@
 #pragma once
 #include "Vehicle.h"
-#include "DrivingSimulator.h"
-/*
+#include "SimulationNode.h"
+#include "SimulationEnvironment.h"
+
 namespace Simulation
 {
 	class AccelerationCalculator
@@ -24,4 +25,3 @@ namespace Simulation
 		simulationNode NextPoint;
 	};
 }
-*/
